micromath/factorial/requirements.txt
Flask~=1.1.2
gunicorn~=20.0.4
jsonschema~=3.2.0
python-logstash-async~=2.1.1
pytest~=6.2.4
coverage~=5.5
